pyproject.toml
[project]
name = "rag-agent-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "nltk>=3.9",
    "sentence-transformers>=5.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.21",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pip>=26.0.1",
    "spacy>=3.8.13",
    "openrouter>=0.8.1",
]
